deleted old api server modules and updated thesandlot server module to take out hardcoded ip and added apifuncs not the actual apis
