+++ thesandlot.py
@@ -1,4 +1,7 @@
-# Starts the web server for the MLB GameDay Retrieval
+#!/usr/bin/env python
 from webapp import webapp
 
-webapp.run(host='127.0.0.1', port=5000, debug=True)
+application = webapp
+
+if __name__ == '__main__':
+    application.run(debug=True)
